+++ app/assets/stylesheets/jikokus.scss
@@ -1,3 +1,3 @@
-// Place all the styles related to the linebot controller here.
+// Place all the styles related to the jikokus controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: https://sass-lang.com/
