--- app/helpers/jikoku_helper.rb
@@ -1,2 +1,0 @@
-module JikokuHelper
-end
