+++ app/helpers/jikokus_helper.rb
@@ -1,0 +1,2 @@
+module JikokusHelper
+end
