--- app/helpers/linebot_helper.rb
@@ -1,2 +1,0 @@
-module LinebotHelper
-end
