+++ app/helpers/linebots_helper.rb
@@ -1,0 +1,2 @@
+module LinebotsHelper
+end
