--- app/views/jikoku/index.html.erb
@@ -1,46 +1,0 @@
-  <body>
-    <div class="d-flex flex-row justify-content-end align-items-center border-bottom shadow-sm">
-      <h5 class="p-4 my-0 mr-md-auto font-weight-normal">
-        Today:
-        <%= l Time.now %>
-        <%= "(#{@current_wday})" %>
-      </h5>
-    </div>
-
-    <div class="pricing-header px-3 py-3 pt-md-5 pb-md-4 mx-auto">
-      <h1 class="display-4 text-center">Time Deference</h1>
-      <p class="lead text-center">このサイトでは、防犯カメラ等の時差を確認することができます。</p>
-    </div>
-
-  <div>
-    <div class="card-header fixed-bottom text-center">
-          <h3 id="RealTimeArea1"class="card-title"></h3>
-    </div>
-    
-    <div class="container" id="card-area-1">
-      <div class="card mb-3  shadow-sm text-center">
-          <div class="card-body">
-            <div class="container-fluid">
-              <h5><label for="InputDateTime">現時表示の時刻を選択</label></h5>
-                <input type="datetime-local" id="InputDateTime" class="shadow-sm" style= "width: 200px;">
-                <input type="number" id="InputSecondTime" class="shadow-sm" placeholder="秒数を入力" style= "width: 120px;" min="0" max="59"><label>秒</label>
-                <button type="button" class="btn-time-deference btn-primary btn-sm" id="check-current-time">確認</button>
-              <h3 id="RealTimeArea1"class="card-title"></h3>
-              <h2 id="TimeDeference"></h2>
-              <h5><label for="InputDateTimePass">再生中の表示時刻を選択</label></h5>
-                <input type="datetime-local" class="shadow-sm" id="InputDateTimePass" style= "width: 200px;">
-                <input id="InputSecondTimePass" type="number" class="shadow-sm" placeholder="秒数を入力" style= "width: 120px;" min="0" max="59"><label>秒</label> 
-              <button type="button" class="btn-correction-time btn-primary btn-sm" id="check-correction-time">確認</button>
-              <h2 id ="CorrectionTime"></h2>
-              <div class="text-center">
-                  <button class="m-2 btn btn-outline-success add-button" style="width: 200px; display: none;" id="add-btn-1">追加</button>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-  </div>
-    </div>
-
-  </body>
-</html>
