+++ spec/requests/jikokus_request_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "Jikokus", type: :request do
+
+end
