+++ spec/requests/linebots_request_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "Linebots", type: :request do
+
+end
